+++ CMakeLists.txt
@@ -237,9 +237,8 @@ list(APPEND NVFUSER_SRCS
   ${NVFUSER_SRCS_DIR}/scheduler/heuristic.cpp
   ${NVFUSER_SRCS_DIR}/scheduler/mark_aliases.cpp
   ${NVFUSER_SRCS_DIR}/scheduler/matmul.cpp
-  ${NVFUSER_SRCS_DIR}/scheduler/multi_matmul.cpp
-  ${NVFUSER_SRCS_DIR}/scheduler/multi_matmul_ampere-.cpp
-  ${NVFUSER_SRCS_DIR}/scheduler/multi_matmul_hopper+.cpp
+  ${NVFUSER_SRCS_DIR}/scheduler/matmul_ampere-.cpp
+  ${NVFUSER_SRCS_DIR}/scheduler/matmul_hopper+.cpp
   ${NVFUSER_SRCS_DIR}/scheduler/matmul_heuristic_plugin.cpp
   ${NVFUSER_SRCS_DIR}/scheduler/matmul_utils.cpp
   ${NVFUSER_SRCS_DIR}/scheduler/mma_utils.cpp
